feat: set swagger to controllers

+++ src/main/java/com/karoldm/k_board_api/controllers/TaskController.java
@@ -3,6 +3,7 @@ package com.karoldm.k_board_api.controllers;
 import com.karoldm.k_board_api.dto.payload.AddResponsiblePayloadDTO;
 import com.karoldm.k_board_api.dto.payload.EditTaskPayloadDTO;
 import com.karoldm.k_board_api.dto.payload.TaskPayloadDTO;
+import com.karoldm.k_board_api.dto.response.ErrorResponseDTO;
 import com.karoldm.k_board_api.dto.response.TaskResponseDTO;
 import com.karoldm.k_board_api.entities.Project;
 import com.karoldm.k_board_api.entities.Task;
@@ -11,6 +12,11 @@ import com.karoldm.k_board_api.mappers.TaskMapper;
 import com.karoldm.k_board_api.services.ProjectService;
 import com.karoldm.k_board_api.services.TaskService;
 import com.karoldm.k_board_api.services.UserService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -30,6 +36,15 @@ public class TaskController {
     private final ProjectService projectService;
 
     @PostMapping
+    @Operation(
+            summary = "create a new task")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "created successfully"),
+            @ApiResponse(responseCode = "401", description = "unauthorized", content = @Content(schema = @Schema(implementation = ErrorResponseDTO.class))),
+            @ApiResponse(responseCode = "400", description = "invalid body data", content = @Content(schema = @Schema(implementation = ErrorResponseDTO.class))),
+            @ApiResponse(responseCode = "403", description = "user is not owner neither member of the project", content = @Content(schema = @Schema(implementation = ErrorResponseDTO.class))),
+
+    })
     public ResponseEntity<TaskResponseDTO> createTask(@RequestBody @Valid TaskPayloadDTO data) {
         checkProjectOwnershipOrParticipation(data.projectId());
         Project project = getProjectOrThrow(data.projectId());
@@ -51,6 +66,13 @@ public class TaskController {
     }
 
     @GetMapping("/{projectId}")
+    @Operation(
+            summary = "get tasks by project")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "get successfully"),
+            @ApiResponse(responseCode = "401", description = "unauthorized", content = @Content(schema = @Schema(implementation = ErrorResponseDTO.class))),
+            @ApiResponse(responseCode = "403", description = "user is not owner neither member of the project", content = @Content(schema = @Schema(implementation = ErrorResponseDTO.class))),
+    })
     public ResponseEntity<Set<TaskResponseDTO>> getAllTasksByProject(@PathVariable UUID projectId, @RequestParam Optional<UUID> memberId) {
         checkProjectOwnershipOrParticipation(projectId);
         Project project = getProjectOrThrow(projectId);
@@ -72,6 +94,14 @@ public class TaskController {
     }
 
     @PutMapping("/{taskId}/member")
+    @Operation(
+            summary = "add a member to a task")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "added successfully"),
+            @ApiResponse(responseCode = "401", description = "unauthorized", content = @Content(schema = @Schema(implementation = ErrorResponseDTO.class))),
+            @ApiResponse(responseCode = "403", description = "user is not owner neither member of the project", content = @Content(schema = @Schema(implementation = ErrorResponseDTO.class))),
+            @ApiResponse(responseCode = "404", description = "task not found", content = @Content(schema = @Schema(implementation = ErrorResponseDTO.class))),
+    })
     public ResponseEntity<TaskResponseDTO> addMember(@PathVariable UUID taskId, @RequestBody @Valid AddResponsiblePayloadDTO data) {
 
         Optional<Task> task = taskService.findTaskById(taskId);
@@ -79,6 +109,8 @@ public class TaskController {
         if(task.isEmpty()){
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Task not found with ID: " + taskId);
         }
+
+        checkProjectOwnershipOrParticipation(task.get().getProject().getId());
 
         Set<UUID> membersId = data.membersId();
         Project project = task.get().getProject();
@@ -98,6 +130,14 @@ public class TaskController {
     }
 
     @PutMapping("/{taskId}")
+    @Operation(
+            summary = "edit task's info")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "edited successfully"),
+            @ApiResponse(responseCode = "401", description = "unauthorized", content = @Content(schema = @Schema(implementation = ErrorResponseDTO.class))),
+            @ApiResponse(responseCode = "403", description = "user is not owner neither member of the project", content = @Content(schema = @Schema(implementation = ErrorResponseDTO.class))),
+            @ApiResponse(responseCode = "404", description = "task not found", content = @Content(schema = @Schema(implementation = ErrorResponseDTO.class))),
+    })
     public ResponseEntity<TaskResponseDTO> editTask(@PathVariable UUID taskId, @RequestBody @Valid EditTaskPayloadDTO data) {
 
         Optional<Task> task = taskService.findTaskById(taskId);
@@ -106,18 +146,30 @@ public class TaskController {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Task not found with ID: " + taskId);
         }
 
+        checkProjectOwnershipOrParticipation(task.get().getProject().getId());
+
         Task updatedTask = taskService.editTask(task.get(), data);
 
         return ResponseEntity.ok(TaskMapper.toTaskResponseDTO(updatedTask));
     }
 
     @DeleteMapping("/{taskId}")
+    @Operation(
+            summary = "delete task")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "deleted successfully"),
+            @ApiResponse(responseCode = "401", description = "unauthorized", content = @Content(schema = @Schema(implementation = ErrorResponseDTO.class))),
+            @ApiResponse(responseCode = "403", description = "user is not owner neither member of the project", content = @Content(schema = @Schema(implementation = ErrorResponseDTO.class))),
+            @ApiResponse(responseCode = "404", description = "task not found", content = @Content(schema = @Schema(implementation = ErrorResponseDTO.class))),
+    })
     public ResponseEntity<Void> deleteTask(@PathVariable UUID taskId) {
         Optional<Task> task = taskService.findTaskById(taskId);
 
         if(task.isEmpty()){
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Task not found with ID: " + taskId);
         }
+
+        checkProjectOwnershipOrParticipation(task.get().getProject().getId());
 
         taskService.deleteTask(taskId);
